Adding ini file to gptp. Issue #115
The same idea used inside the avtp_pipeline was used here.
We are now adding the possibility to use a configuration (ini) file
to configure the internal variables of gptp.

Currently the following was added:
 - 3rd party ini file parser;
 - Class with callback to the parser
 - some options were added as an example to the gptp_cfg.ini file
 - all the current options are parsed inside linux/daemon_cl.cpp and printed out during init.

Still missing:
 - use the options from the file inside gPTP

Also, the linux daemon_cl.cpp was indented since there were several indenting problems
on the file.