64 bit words network/host order conversion
When using 64 bit words, the function byte_swap64 is used. This
function assumes the machine's endianness and because of that it is
not portable.

This patch adds two new functions to deal with the problem of converting
64 bit words between host/network order without assuming the host endianness.

Two byte_swap64 calls were replaced by these two new functions in
PTPMessageCommon::buildCommonHeader and
PTPMessageCommon::buildPTPMessage.

This commit addresses Issue #243