gPTP: Fix rate ratio calculation
From the 802.1AS-2011 standard (clause 11.2.13.2), rate ratio is
defined as ( cumulativeScaledRateOffset * 2^(-41) ) + 1. Which is
equivalent to ( cumulativeScaledRateOffset / 2^(41) ) + 1.

But in the calculation for master_local_freq_offset in ptp_message.cpp
line 961; the cumulativeScaledRateOffset is multiplied by 2ULL << 41
which effectively is 2^42.

Fix the calculation of the exponent by changing it to 1ULL << 41.