Fix for gPTP interval init on ARM platforms
PortInit_t used char fields to store signed sentinel values to
indicate that default values should be set by the port. This was
fine for Intel, which generally has a signed representation for
the char type, but the sentinel values were not recognized on ARM
platforms, where char is generally unsigned.

Changing the fields to explicitly be of signed char type resolves
the problem on ARM and keeps the same behavior on Intel.