offsetScaledLogVariance: change it to unsigned short (SCA _032 #65)
Static code analysis fix _032 (Issue #65)

SCA complains here about ntohs() but the real issue is the type of offsetScaledLogVariance,
which currently in the code is int16_t. There seems to be inconsistency in IEEE 802.1AS-2011:

    - in 6.3.3.8 ClockQuality offsetScaledLogVariance is defined as UInteger16;
    - later in the document it's used as Integer16 (Table 14-1, code examples etc.);

This inconsistency has been corrected in IEEE 802.1AS-2011/Cor 1-2013 where UInteger16 was specified
in the Table 14-1 and used later in the code examples.